[SERVICE-508] Reorganise test files
