+++ test/demo/contextChannels/broadcastOnWindow.inttest.ts
@@ -1,10 +1,9 @@
 import 'jest';
 import {connect, Fin, Identity, Application} from 'hadouken-js-adapter';
 
-import {ChannelId} from '../../src/client/contextChannels';
-
-import * as fdc3Remote from './utils/fdc3RemoteExecution';
-import {appStartupTime} from './constants';
+import {ChannelId} from '../../../src/client/contextChannels';
+import * as fdc3Remote from '../utils/fdc3RemoteExecution';
+import {appStartupTime} from '../constants';
 
 /*
  * Tests how existing broadcast() and addContextListener() calls interact with Channel.join()
@@ -19,7 +18,7 @@ const startedApps:Application[] = [];
 let fin: Fin;
 
 beforeAll(async () => {
-    fin = await connect({address: `ws://localhost:${process.env.OF_PORT}`, uuid: 'TEST-contextChannelsWindowBroadcasting.ts'});
+    fin = await connect({address: `ws://localhost:${process.env.OF_PORT}`, uuid: 'TEST-contextChannels-broadcastOnWindow.inttest.ts'});
     await expect(fin.Application.wrapSync({uuid: 'test-app', name: 'test-app'}).isRunning()).resolves.toBe(true);
 });
