+++ test/demo/contextChannels/get.inttest.ts
@@ -1,10 +1,9 @@
 import {connect, Fin} from 'hadouken-js-adapter';
 import {Identity} from 'openfin/_v2/main';
 
-import {DesktopChannel, DefaultChannel, ChannelError, IdentityError} from '../../src/client/main';
-
-import * as fdc3Remote from './utils/fdc3RemoteExecution';
-import {testManagerIdentity, testAppInDirectory1, testAppNotInDirectory, appStartupTime, testAppNotFdc3} from './constants';
+import {DesktopChannel, DefaultChannel, ChannelError, IdentityError} from '../../../src/client/main';
+import * as fdc3Remote from '../utils/fdc3RemoteExecution';
+import {testManagerIdentity, testAppInDirectory1, testAppNotInDirectory, appStartupTime, testAppNotFdc3} from '../constants';
 
 /**
  * Tests getDesktopChannels(), getChannelById(), and getCurrentChannel()
@@ -13,7 +12,7 @@ import {testManagerIdentity, testAppInDirectory1, testAppNotInDirectory, appStar
 let fin: Fin;
 
 beforeAll(async () => {
-    fin = await connect({address: `ws://localhost:${process.env.OF_PORT}`, uuid: 'TEST-contextChannelQuerying.ts'});
+    fin = await connect({address: `ws://localhost:${process.env.OF_PORT}`, uuid: 'TEST-contextChannels-get.inttest.ts'});
     await expect(fin.Application.wrapSync({uuid: 'test-app', name: 'test-app'}).isRunning()).resolves.toBe(true);
 });
