+++ test/demo/contextChannels/join.inttest.ts
@@ -1,11 +1,10 @@
 import {connect, Fin} from 'hadouken-js-adapter';
 import {Identity} from 'openfin/_v2/main';
 
-import {IdentityError} from '../../src/client/main';
-
-import {testManagerIdentity, appStartupTime, testAppNotInDirectory, testAppNotFdc3, testAppInDirectory1} from './constants';
-import * as fdc3Remote from './utils/fdc3RemoteExecution';
-import {RemoteChannel} from './utils/RemoteChannel';
+import {IdentityError} from '../../../src/client/main';
+import {testManagerIdentity, appStartupTime, testAppNotInDirectory, testAppNotFdc3, testAppInDirectory1} from '../constants';
+import * as fdc3Remote from '../utils/fdc3RemoteExecution';
+import {RemoteChannel} from '../utils/RemoteChannel';
 
 /*
  * Tests simple behaviour of Channel.getMembers() and the channel-changed event, before testing how they and getCurrentChannel()
@@ -15,7 +14,7 @@ import {RemoteChannel} from './utils/RemoteChannel';
 let fin: Fin;
 
 beforeAll(async () => {
-    fin = await connect({address: `ws://localhost:${process.env.OF_PORT}`, uuid: 'TEST-contextChannelJoining.ts'});
+    fin = await connect({address: `ws://localhost:${process.env.OF_PORT}`, uuid: 'TEST-contextChannels-join.ts'});
     await expect(fin.Application.wrapSync({uuid: 'test-app', name: 'test-app'}).isRunning()).resolves.toBe(true);
 });
